prettify code. fix rubocop waning. move all methods with change log generation to generator
