+++ .rubocop_todo.yml
@@ -1,5 +1,5 @@
 # This configuration was generated by `rubocop --auto-gen-config`
-# on 2015-05-22 14:05:01 +0300 using RuboCop version 0.31.0.
+# on 2015-05-22 15:54:38 +0300 using RuboCop version 0.31.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -16,7 +16,7 @@ Metrics/BlockNesting:
 # Offense count: 3
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 321
+  Max: 294
 
 # Offense count: 5
 Metrics/CyclomaticComplexity:
@@ -47,16 +47,11 @@ Style/AndOr:
 Style/BlockDelimiters:
   Enabled: false
 
-# Offense count: 4
+# Offense count: 6
 Style/Documentation:
   Enabled: false
 
-# Offense count: 2
-# Cop supports --auto-correct.
-Style/EmptyLines:
-  Enabled: false
-
-# Offense count: 1
+# Offense count: 4
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, SupportedStyles.
 Style/EmptyLinesAroundClassBody:
